+++ fetch_data/sparql_data/pull_wikidata.py
@@ -7,7 +7,7 @@ import os
 import sys
 from SPARQLWrapper import QueryResult, SPARQLWrapper, JSON
 
-from logger.get_logger import logger
+from logger import logger
 
 
 async def get_sparql_results(endpoint_url: str, query: str, query_name: str) -> dict:
