+++ logger/__init__.py
@@ -1,0 +1,3 @@
+"""Logger"""
+
+from logger.get_logger import logger
