+++ logger/extra_formatter.py
@@ -1,0 +1,51 @@
+"""Log Format"""
+
+import logging
+
+
+DEFAULT_ARGS = {
+    "name",
+    "msg",
+    "args",
+    "levelname",
+    "levelno",
+    "pathname",
+    "filename",
+    "module",
+    "exc_info",
+    "exc_text",
+    "stack_info",
+    "lineno",
+    "funcName",
+    "created",
+    "msecs",
+    "relativeCreated",
+    "thread",
+    "threadName",
+    "processName",
+    "process",
+    "wikibase",
+}
+
+
+class OptionalExtraFormatter(logging.Formatter):
+    """Supports an optional `wikibase` extra argument"""
+
+    def format(self, record: logging.LogRecord) -> str:
+        """Format Record"""
+
+        base_fmt = "%(asctime)s | %(levelname)s | %(message)s"
+        wikibase_fmt = (
+            "%(asctime)s | %(levelname)s | wikibase %(wikibase)d | %(message)s"
+        )
+
+        if "wikibase" in record.__dict__.keys():
+            self._style._fmt = wikibase_fmt  # pylint: disable=protected-access
+        else:
+            self._style._fmt = base_fmt  # pylint: disable=protected-access
+
+        for k in record.__dict__.keys():
+            if k not in DEFAULT_ARGS:
+                self._style._fmt += f"\n\t{k}: {record.__dict__.get(k)}"
+
+        return super().format(record)
