+++ model/enum/wikibase_log_type_enum.py
@@ -6,7 +6,7 @@ from typing import Optional
 
 import strawberry
 
-from logger.get_logger import logger
+from logger import logger
 
 
 @strawberry.enum
